third_party/WebKit/LayoutTests/webaudio/PeriodicWave/periodicwave-contexts-expected.txt

CONSOLE WARNING: line 101: Oscillator.frequency.exponentialRampToValue value 24050 outside nominal range [-22050, 22050]; value will be clamped.
Test Builtin Oscillator PeriodicWave Tables

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 80 dB
PASS Maximum difference below threshold of 4.06 ulp (16-bits)
PASS Number of differences between actual and expected result is less than 15827 out of 176400
PASS successfullyParsed is true

TEST COMPLETE
